Completed Code. Need to add comments and tests

+++ src/AmbientDevice.java
@@ -37,10 +37,10 @@ public class AmbientDevice {
      * Setter for the brightness of the lamp with a value between 0.0-100.0
      * @param value Brightness of lamp
      */
-    public void setBrightness(float value) {
-        if (value > 100.0f) value = 100.0f;
-        if (value < 0.0f) value = 0.0f;
-        this.brightness = value / 100;
+    public void setBrightness(int value) {
+        if (value > 100) value = 100;
+        if (value < 0) value = 0;
+        this.brightness = ((float)value / 100);
         lampColor = Color.getHSBColor(hue, saturation, this.brightness);
         lamp.setColor(lampColor);
     }
@@ -49,10 +49,10 @@ public class AmbientDevice {
      * Setter for the color of the lamp with a value between 0.0-100.0
      * @param value Color of lamp
      */
-    public void setColor(float value) {
-        if (value > 100.0f) value = 100.0f;
-        if (value < 0.0f) value = 0.0f;
-        this.hue = value / 100;
+    public void setColor(int value) {
+        if (value > 100) value = 100;
+        if (value < 0) value = 0;
+        this.hue = ((float)value / 100);
         lampColor = Color.getHSBColor(this.hue, saturation, brightness);
         lamp.setColor(lampColor);
     }
